update the dependencies
