+++ src/components/notification/alert-message.ts
@@ -29,7 +29,7 @@ import * as Handlebars from 'handlebars'
 import getos from 'getos'
 import osName from 'os-name'
 import { getContext } from '../../context'
-import type { NotificationMessage } from '../notification/channel'
+import { NotificationMessage } from '@hyperjumptech/monika-notification/channel'
 import { ProbeRequestResponse } from '../../interfaces/request'
 import { ProbeAlert } from '../../interfaces/probe'
 import { publicIpAddress, publicNetworkInfo } from '../../utils/public-ip'
